fixed public facing views, fixed server model issue
