+++ modules/events/client/config/events-admin.client.config.js
