+++ modules/events/client/config/events-admin.client.menus.js
