+++ modules/events/client/controllers/admin/add-event.client.controller.js
