+++ modules/events/client/services/events-admin.client.service.js
@@ -1,0 +1,13 @@
+'use strict';
+
+angular.module('events.admin').factory('eventsAdmin', ['$resource',
+  function ($resource) {
+      return $resource('api/events/:eventId', {
+          eventId: '@_id'
+      }, {
+          update: {
+              method: 'PUT'
+          }
+      });
+  }
+]);
