+++ modules/events/client/services/events.client.service.js
@@ -1,13 +1,9 @@
 'use strict';
 
-angular.module('events.admin').factory('eventsAdmin', ['$resource',
+angular.module('events').factory('eventsPublic', ['$resource',
   function ($resource) {
       return $resource('api/events/:eventId', {
           eventId: '@_id'
-      }, {
-          update: {
-              method: 'PUT'
-          }
       });
   }
 ]);
