+++ modules/events/client/views/list-events.client.view.html
@@ -1,0 +1,21 @@
+<section>
+  <div class="page-header">
+    <div class="row">
+      <div class="col-md-4">
+        <h1>Events</h1>
+      </div>
+      <div class="col-md-4" style="margin-top: 2em">
+        <input class="form-control col-md-4" type="text" ng-model="search" placeholder="Search" ng-change="figureOutItemsToDisplay()" />
+      </div>
+    </div>
+  </div>
+  <div class="list-group">
+    <a ng-repeat="event in pagedItems" ui-sref="event({eventId: event._id})" class="list-group-item">
+      <h4 class="list-group-item-heading" ng-bind="event.title"></h4>
+      <p class="list-group-item-text pull-right small" ng-bind="event.image"></p>
+      <p class="list-group-item-text" ng-bind="event.description"></p>
+    </a>
+  </div>
+  <pagination boundary-links="true" max-size="8" items-per-page="itemsPerPage" total-items="filterLength" ng-model="currentPage" ng-change="pageChanged()"></pagination>
+
+</section>
