+++ modules/events/client/views/view-event.client.view.html
@@ -1,0 +1,34 @@
+﻿<section>
+    <div class="page-header">
+        <div class="row">
+            <div class="col-md-6">
+                <h1 ng-bind="event.title"></h1>
+            </div>
+        </div>
+    </div>
+
+    <div class="row">
+        <div class="col-md-8">
+            <div class="row">
+                <div class="col-md-3"><strong>Title</strong></div>
+                <div class="col-md-6" ng-bind="event.title"></div>
+            </div>
+            <hr />
+            <div class="row">
+                <div class="col-md-3"><strong>Description</strong></div>
+                <div class="col-md-6" ng-bind="event.description"></div>
+            </div>
+            <hr />
+            <div class="row">
+                <div class="col-md-3"><strong>image</strong></div>
+                <div class="col-md-6" ng-bind="event.image"></div>
+            </div>
+            <hr />
+            <div class="row">
+                <div class="col-md-3"><strong>Created</strong></div>
+                <div class="col-md-6" ng-bind="event.created_at"></div>
+            </div>
+            <hr />
+        </div>
+    </div>
+</section>
